[rpc] Refactor - expose RPC functions in rpc package
