+++ rpc/apiv1/error.go
