+++ rpc/apiv2/error.go
