+++ rpc/apiv2/harmony.go
@@ -5,8 +5,8 @@ import (
 	"math/big"
 
 	"github.com/harmony-one/harmony/hmy"
-	commonRPC "github.com/harmony-one/harmony/internal/hmyapi/common"
 	"github.com/harmony-one/harmony/internal/params"
+	"github.com/harmony-one/harmony/rpc/common"
 )
 
 // PublicHarmonyAPI provides an API to access Harmony related information.
@@ -59,11 +59,11 @@ type NodeMetadata struct {
 }
 
 // GetNodeMetadata produces a NodeMetadata record, data is from the answering RPC node
-func (s *PublicHarmonyAPI) GetNodeMetadata() commonRPC.NodeMetadata {
+func (s *PublicHarmonyAPI) GetNodeMetadata() common.NodeMetadata {
 	return s.hmy.GetNodeMetadata()
 }
 
 // GetPeerInfo produces a NodePeerInfo record, containing peer info of the node
-func (s *PublicHarmonyAPI) GetPeerInfo() commonRPC.NodePeerInfo {
+func (s *PublicHarmonyAPI) GetPeerInfo() common.NodePeerInfo {
 	return s.hmy.GetPeerInfo()
 }
